refactor: extract header functionalities into separate file
